+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Merged pull requests:**
 
+- Add Tracing Tags Variables [\#7](https://github.com/luke-taylor/terraform-azurerm-management/pull/7) ([luke-taylor](https://github.com/luke-taylor))
 - Update Log Analytics Solutions Variable  [\#6](https://github.com/luke-taylor/terraform-azurerm-management/pull/6) ([luke-taylor](https://github.com/luke-taylor))
 - Remove Unused Data Block and Clean-up [\#5](https://github.com/luke-taylor/terraform-azurerm-management/pull/5) ([luke-taylor](https://github.com/luke-taylor))
 - Bump github.com/gruntwork-io/terratest from 0.41.23 to 0.41.24 in /test [\#4](https://github.com/luke-taylor/terraform-azurerm-management/pull/4) ([dependabot[bot]](https://github.com/apps/dependabot))